refactor(water): split BGSWaterCollisionManager.h per class
One RTTI'd class per file matches repo convention (e.g. BSResourceEntry.h/
BSResourceHandle.h split despite sharing the BSResource namespace).
